backend/requirements.txt
# Lightweight FastAPI Backend for Render deployment
# Excludes heavy ML dependencies (torch, transformers, sentence-transformers)

# Core FastAPI
fastapi==0.123.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# API integrations
openai==1.24.0
google-generativeai==0.5.2
qdrant-client==1.9.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Authentication
python-jose[cryptography]==3.4.0
bcrypt==3.2.0
python-multipart==0.0.18

# HTTP & utilities
httpx==0.26.0
tiktoken==0.7.0

# Lightweight RAG dependencies
langchain==0.1.20
langchain-openai==0.1.7
langchain-community==0.0.38

# Skip these heavy packages for Render:
# torch==2.3.0 (2GB+)
# transformers==4.40.0 (requires torch)
# sentence-transformers==2.7.0 (requires torch)
# faiss-cpu==1.8.0 (heavy, use Qdrant instead)
# chromadb==0.4.24 (heavy, use Qdrant instead)
